Adds `tsconfig.json`.
Generated with `npx typescript@3.9 --init`. There were two subsequent edits to the file:
1. Updated `target` and `lib` to be in line with Node 12.13, which is used by default by Bazel.
2. Excluded `bazel-*/` and `dist/` or else VSCode will not be able to load properly, likely due to an overwheling amount of files to check.

Also exported the file from the root `BUILD.bazel` file, so `ts_library()` rules can use it.

Finally, I referenced this in the VSCode settings config, so the same version of TypeScript should be used in the VSCode language service.