Imports and invokes the given entry point.
This uses a dynamic `import()` to load the provided entry point JavaScript and invoke its default export function.

Node v12.13 does not yet support ES modules, so `import()` actually gets transpiled to `require()`. There are some ways to mock it, but I figured the simplest and easiest was to just make a wrapper function that could be trivially spied on. This allows us to test all the error cases.

For now it just invokes the default export and prints the results, validating everything along the way.