Refactors `renderer` build to export a library.
While working on the Bazel rules, I found that it is impossible to dynamically import a file that is not an explicit dependency of the `nodejs_binary()`. With the old design, the renderer was a `nodejs_binary()` which would accept a JavaScript file an input, dynamically importing it. However, `nodejs_binary()` adds a wrapper for `require()` statements which provides path mapping to the binary's runfiles directory. Since user code would be an input to the `nodejs_binary()`, not a dependency, it would not be present in runfiles. This means that the `require()` wrapper cannot import the user-provided module. It is possible to work around this by importing an absolute path, however if that file has its own dependencies, then those will **also** need to use absolute imports, and there is no good tooling to make that happen.

To work around this, I'm rearchitecting the `renderer` slightly. Instead of building a single `nodejs_binary()` to use for all `prerender_page()` rules, each `prerender_page()` will generate its own `nodejs_binary()` tool. The actual implementation is mostly the same, we will still dynamically import the user-provided module as specified via the `--entry-point` flag. However, `prerender_page()` will be responsible for creating the `nodejs_binary()` which includes the user-provided module and its dependencies as a `data` dep. This means that all the user code will be present in the runfiles directory where it can be imported as the Bazel gods intended.

Practically speaking, this means that `//renderer/...` no longer exports `nodejs_binary()`, but rather a `ts_library()` and a set of JavaScript sources, which will be needed to construct a `nodejs_binary()` with the proper entry point. We still build a `nodejs_binary()` for testing purposes, but that should never be used in a real build.