Adds `prerender_page()` rule and macro.
This allows users to easily compile their own TypeScript generator and invoke it at build time. Example to follow.

The `prerender_page()` macro generates a new `ts_library()` with user input. We could have simply accepted this as a label to give the user more control, however this will be necessary in the future to provide strict dependencies. Otherwise a user could depend on a component's render code without also depending on the client JavaScript and CSS, which is the whole point of `rules_prerender`.

We have to generate a new `nodejs_binary()` for each `prerender_page()` so that it can depend on the user-provided render code. This needs to be a dependency or else the renderer will not be able to import it due to restrictions in the `require()` wrapper (see the previous commit message for details).