Adds minimal example use of `prerender_page()`.
This renders a simple "Hello, World!" page. Imports a direct dependency and uses a transitive dependency just to make sure all the infra is working correctly.

Also includes a golden test by simply diffing the file with an existing golden. Unfortunately VSCode keeps adding a trailing newline which was ruining the comparison. I tried mangling VSCode settings to make an exception for golden files, but it didn't work for reasons I can't explain. Instead, the easiest solution is to just add a trailing newline to the actual implementation. A little awkward, but at least means I'm not fighting the editor.