Adds a devserver for the minimal example.
I utilized `ts_devserver()` as a Bazel-compatible out-of-the-box implementation of a devserver, although currently it is only serving the one HTML file.

I renamed the generated HTML file to `index.html` so it is used as a home page and is more easily accessible for unfamiliar devs.

I also added a JS script to the HTML which is generated by the devserver. This implements live reload function so `ibazel run //minimal/example:devserver` "just works" with automatic builds and live-reload.