Adds `delegatedProxy()`.
This is to be used for Jasmine effects to couple `beforeEach()` and `afterEach()` functionality in a convenient wrapper. The proxy serves to represent the result, dynamically delegating individual operations to the initialized value that came from the `beforeEach()`.

I tried to implement all the possible functions on the `ProxyHandler`, but I had to skip a few. I skipped `enumerate()` because while it does exist in the type system, it is apparently not supported anywhere except old Firefox: https://developer.mozilla.org/en-US/docs/Archive/Web/JavaScript/handler.enumerate

However, object construction and freezing cause some challenges. They have specific invariants which relate the `Proxy` and its target. These invariants are actually validated at runtime and throw errors if not properly followed. Unfortunately I'm using a dummy target here because it is dynamically retrieved when accessed, meaning I'm not able to reflect some of the invariants. For instance, to construct a `Proxy`, its target must be constructable as well, however using a `class {}` or `function () {}` introduces a `prototype` key not normally accessible on most objects and breaks other invariants. I can't dynamically choose what the target is either because I don't know the target until after the proxy is created and the user accesses the `Proxy` for the first time. Freezing the object had similar problems, where the target must have the same properties as the proxy, which is very difficult to enforce. Instead, I just left these features unsupported.