Adds `effects` utilities.
These aid the creation of what I am dubbing "effects" because I haven't seen this pattern before, I have no imagination, and they kind of look like React hooks if you squint at them. These effects encapsulate the setup and cleanup code for a particular resource. This allows tests using such a resource to just call the effect and not worry about initializing and cleaning up the resource between tests. No more awkward `let` declarations, non-nullable types that are actually nullable, or forgetting to add an `afterEach()` cleanup!

The secret sauce that makes this work is that an effect calls `beforeEach()` and `afterEach()` (or `beforeAll()` / `afterAll()`) which contains the initialization and cleanup code together. However, it can't actually return the resource, because the initialization is in `beforeEach()`, which hasn't been called yet at the time the effect is invoked. To work around this, we instead return a `Proxy`, which proxies all operations to the real value after its initialization. This allows the `Proxy` to exist before initialization and after cleanup, working nicely with `const` and strict null checks, but still functioning like the actual resource it represents.

There are a few caveats to be aware of:
1. The resource is only usable between initialization and cleanup, even though it can be referenced outside those bounds. Well-behaved tests won't do this, so it's unlikely to be a significant issue, but could definitely arise from edge cases.
2. This is using `delegatedProxy()`, which has a few limitations compared to a typical `Proxy` object (see its own documentation for what those are).
3. The differing internal structure of `Proxy` objects can make them behave is subtly different ways. For instance, using `expect(resource).toEqual({ /* ... */ })` is likely to not work because the internal representation is subtly different. Effects are intended for test resources, so you generally shouldn't be asserting on the whole object like that anyways.

I haven't seen this kind of pattern used before, so I have no idea if this is actually a good idea. As of right now it is only an "interesting" idea, so I'm going to try it out just to see where it goes.