Adds `devserver` test utilities.
This provides some wrapper functionality to start a child `ts_devserver()` process and run it in the background for use in tests. This can serve as a simple static file server implementation to host a site visited a test-controlled Chrome instance.

The devserver can be used directly with `Server.spawn()` and `Server.prototype.kill()` or via the `useDevserver()` effect which manages that for the test. I decided to use `beforeAll()` and `afterAll()` as opposed to `beforeEach()` and `afterEach()` since the server should be the same for all tests executing and should be relatively stable. The performance impact of starting up and shutting down a new server with each test would likely not be worth the potential isolation benefit.

Running the server is a bit tricky as it isn't really designed for programmatic use cases like this. I identified a port by listening to port `0`, which the OS interprets as "Give me any open port". Then the test process releases that port and invokes the devserver binary with it. This isn't ideal, as potentially another process could steal the port in that time period, but within the context of a test inside the Bazel sandbox, I'm not too concerned about the issue.

Killing the server is also awkward, since it is intended to be killed via Ctrl+C from a user. I simulated this by sending the `SIGINT` signal, although that alone isn't enough, as the signal actually needs to be sent to all processes in the tree of the devserver process, so I loaded the `tree-kill` utility to handle that. It's unfortunate that this is necessary, but not that big a deal in the end.