Replaces golden test with a Puppeteer test.
This removes the existing golden test setup and replaces it with a sparkly new Puppeteer-based test. It runs the `ts_devserver()`, starts Chrome, visits the page, and asserts on the elements. This is quite a bit overkill based on the currently very simplistic implementation of prerendering, but this is useful infrastructure to have for tests of more complicated use cases going forward. This particular test is very straightforward, just reading the text content of relevant HTML elements. It leverages effects for the devserver and Puppeteer and they "just work" without any additional trickyness.

The trickiest part was understanding how to debug Puppeteer tests. I happen to be using Windows 10 with WSL 2, which makes things even more complicated. Eventually I discovered that it is possible run an Xserver on Windows (not provided by Microsoft by default, but implementations exist on the web) and then proxy that through WSL and pass it into the Bazel test. This is all done with the `--config debug` and documented in the README. The VSCode debugger works reasonably well, since I am able to map back to the source locations, so adding a breakpoint anywhere will just be hit, rather than `chrome://inspect`, which doesn't have access to the source files until after it has run, meaning a `debugger;` statement is needed to really accomplish anything.

See [this answer](https://stackoverflow.com/a/61110604) for more details about setting up Xserver with WSL 2.