Adds `EffectTester`.
This is a test utility for verifying that effects function as expected. Without this, testing effects is quite difficult because they are inherently coupled to Jasmine's `before*()` and `after*()` functions. A lot of bookkeeping needs to be done just to be able to confidently trigger an effect's initialization and cleanup, even before verifying whether those operations actually did what they were expected to do.

`EffectTester` spies on the `before*()` and `after*()` functions and invokes the effect, allowing it to receiving the initialization and cleanup hook registration. It then exposes `initialize()` and `cleanup()` functions which directly invoke the effect to make testing easy. I added a number of additional assertions to verify that the effect is behaving as expected. I also opted **not** to support Jasmine's `done()` callbacks, as those are generally out of fashion anyways and the use case is much better supported by `Promises`.