Adds Puppeteer testing utilities.
This creates effects for using Puppeteer in tests. These are simply responsible for creating/destorying `Browser` and `Page` instances in a developer-friendly fashion.

It also provides a hook to set the `--headless` flag via the `$DISPLAY` environment variable. This is to improve test debugging, where a user can pass `--test_env=DISPLAY` to pass through their `$DISPLAY` variable. This will cause Puppeteer to run in non-headless mode where the Chrome window will be visible to the user and allow them to inspect and debug the page under test in a convenient fashion.

I also added a timeout just shy of Jasmine's timeout. This is to make debugging a little simpler, as blocking on a breakpoint will quickly consume the time limit, and Chrome waits longer the Jasmine. If there is an issue, such as an Xserver connection failure, having this timeout means that Chrome will give its timeout message and draw the user into considering the problem there, rather than with Jasmine.