Adds `prerender_component()` macro.
This provides an encapsulation of "components" for prerendering. A "component" is essentially a grouping of HTML, JS, and CSS to provide a piece of UI to a user. Components can compose one another to build more complex interfaces and eventually generate an entire page. For now, this just includes the prerendered TypeScript portion, no client-side JavaScript or CSS just yet.

This is just done by generating a `ts_library()` for the component. It generates at `%{name}_prerender` specifically to be different from `name` so users don't accidentally depend on it from a `ts_library()`, which would break the dependency management. This also includes a simple example of a few components composing each other with a `build_test()`.