Adds DOM types to `tsconfig.json`.
Puppeteer was typing most values as `any` or `ElementHandle<any>`, requiring awkward type assertions. After some digging, I discovered that several APIs use types along the lines of `T extends Element ? ElementHandle<T> : T` but were being resolved to `any`. Tracing this back, I discovered that the `Element` type was being treated as `any`, because the DOM types are not included in the compilation. Adding them gave a meaningful value for `Element` and made type inference for Puppeteer actually usable.

Unfortunately, this has the side effect that all DOM types are available in all TypeScript files. Considering most of these are NodeJS, this is actually inaccruate, since most DOM types (like `document` for instance) are not actually available in Node and would throw a runtime error if they were ever used. We get around this problem for the built-in Node types because there is an `@npm//@types/node` which selectively opts-in to these types. However, there is no equivalent for the DOM types, so we need to overcorrect and enable them everywhere. There are possible other solutions around multiple `tsconfig.json` files or more Bazel-aware tooling, but I think the more we tweak this the more the IDE and the build system will diverge and only cause more problems. The easiest solution is to just add DOM types everywhere in the repo.