Adds `deps` to `prerender_page()`.
This allows a `prerender_page()` target to depend upon `prerender_component()` targets. All it really does is plumb through the `%{name}_prerender` `ts_library()` targets generated by `prerender_component()` and add them as `deps` of the page generator `ts_library()`.

This also adds a page to the components example with an integration test to verify that it was rendered correctly.