Adds Jasmine wrapper tool.
I recently discovered that forgetting to add a `data` dep on `@npm//puppeteer` for a `jasmine_node_test()` caused the test to pass without actually running anything. Eventually, I traced this behavior back to the `require()` for that dependency which surfaces as an unhandled rejection due to Bazel's Node loader. This rejection then gets ignored by Node and no tests are executed, causing Jasmine to pass. Adding this flag to NodeJS makes it fail if a `Promise` rejection is not handled, effectively catching this error.

This flag is added automatically via a new `//tools:jasmine.bzl` file which wraps `jasmine_node_test()`. Unfortunately, I'm not aware of a good way to enforce that this file is used over `@npm//@bazel/jasmine:index.bzl`, but this is good-enough for now.