Adds default `jasmine.json` configuration file.
This includes the one critical flag, `failSpecWithNoExpectations`, which will fail any `it()` which does not execute an expectation, indicating a broken test. If this behavior is ever desired, such as for a test where an operation simply should complete, then `expect().nothing()` can be used. Also consider `expect(() => /* ... */).not.toThrow()` or `expect(/* ... */).toBeResolved()` if we only care than an error was not thrown.