Adds a `scripts` attribute to `prerender_page()`, allowing users to specify client-side JavaScript libraries to include the prerendered page.
For now, this doesn't really do much as it merely aggregates the libraries under a single target. Users must manually include a `<script />` tag, create an entry point which side-effectfully imports all the files, and then bundle them into a single file with something like Rollup. It will be expanded with the entry point automatically generated and with utilities to inject the `<script />` tag, including tree shaking unused scripts.