Use absolute labels to find sibling targets.
The current implementation does not support using labels with default target names, such as `//foo`, because we use simple string concatenation. This is fixed by converting the target to an absolute label, which expands it to `//foo:foo`, making string concatenation work as expected.

Exposed this as a helper utility so it can be used whereever it is useful.