Adds `prerender_annotation` library.
This allows creating an HTML comment for use as an annotation to prerender build tooling. This can be used to express external resources linked within an HTML tree.

Currently, this only supports JavaScript with a provided path argument. Including this content in an HTML template allows these templates to tree shake scripts, as any JavaScript not rendered into the final page can be dropped from the bundle altogether.

The rendered text is just an HTML comment with a giant `DO_NOT_DEPEND_OR_ELSE` note so users are clear that this is an implementation detail of `rules_prerender` and not something to be used directly. The actual meaningful contents of the comment are expressed as JSON so as to be easily extensible and modifiable going forward.