Adds `includeScript()` function and creates a `rules_prerender` package.
The `includeScript()` function simply creates an annotation for including JavaScript for use by prerender tooling.

This `rules_prerender` package acts as the NPM package that will be installed by users and imported with the `rules_prerender` module. This is done via the `module_name` attribute in the `ts_library()` rule and also by adding it as a path in the root `tsconfig.json` so editors understand where the module is coming from.