Adds `extractor` library.
This parses an HTML string, identifies all the annotations, removes those comments, and then returns the new content with the parsed annotations.

There are two tricky parts here. First is removing the annotation comments from the document. The `remove()` API does not exist for comment nodes, so instead we need to call `removeChild()` on the parent, which requires a lot more bookkeeping than we should really have to do. The second tricky part is deduplicating annotations. We want to deduplicate the parsed annotation so minor string edits don't introduce semantically equivalent elements. This required a deep equals implementation which is historically painful to do in JavaScript. Fortunately, there is an existing Node API I was able to leverage to minimize the maintenance overhead.

I actually tried using the `jsdom` library initially, however it seems to have a bug in parsing child nodes of the body element, where I was not able to read comments under the `<body />`. `node-html-parser` seems like a lighter weight alternative anyways which is powerful enough for the requirements here.