Implements annotation extractor binary.
All this does is call through to the existing libraries for extracting annotations and assembling them into a `PrerenderMetadata`. The binary itself is just responsible for reading and writing the files while these libraries perform all the actual business logic.

Updated the test with a single end-to-end use case. The edge cases are tested in the individual libraries and don't need to be exercised here.