Adds annotation extractor to `prerender_page()` macro.
This means annotations are removed from the HTML before they are returned to the user. It also generates the metadata JSON file which is currently unused. I added a golden test for now to verify that it is working, but otherwise it doesn't actually do anything. Needed to add a trailing newline to the extractor to make the golden test work easily and not fight with VSCode.