Makes all client-side scripts ES modules.
By chance, these scripts do not `import` or `export` anything and as such, TypeScript treats them as non-modules, meaning they share the same namespace. Giving functions the same name causes VSCode to show an error because it believes they conflict at the global scope.

Our build process `import`s these files and bundles them with Rollup, so this isn't a real error. Adding `export {};` is enough for TypeScript to recognize the file as a module and existing in a separate scope.