Adds `entry_generator` package.
This will be used to generate a TypeScript entry point for all the scripts included in a page via annotations. It accepts a metadata file containing a list of scripts and outputs a TypeScript source file which imports all of them. Useful for tools like Rollup which expect explicit imports to all directly included files.