Adds generated entry point to `prerender_page()`.
The macro now generates and includes an entry point which directly imports all the scripts used by the page, referenced by `includeScript()`. This can be plugged directly into `rollup_bundle()` without requiring any custom glue code.

With this, we can delete the `entry.ts` file, as well as the tests of `metadata.json`, which is now used to generate this entry point.

The one notable change is that `includeScript()` actually gets resolved through Node resolution logic. `rollup_bundle()` also uses ES6 sources, so technically the `.mjs` extension should be used. Since the paths are resolved, extensions can be dropped altogether and load correctly, so that is the recommended solution.