Adds `injector` library.
This is responsible for parsing an HTML document and inserting `<script />` tags as specified by an input configuration object. This is the core implementation of the resource injector tool.

The actual implementation is farily straightforward, though some effort was made to scale up as more resources become supported in the future (such as CSS).

I did make an attempt to get things to indent correctly, but the parser API is not conducive to supporting that. There's no easy way to know what the user's tab size is and there is no easy way to delete excess spacing. As a result, I opted to mostly ignore spacing.