Adds `inject_resources()` macro.
This creates a config JSON file with the user provided scripts and passes it with the input HTML into the resource injector binary.

The only tricky part here is encoding the user's configuration as JSON. Bazel does define a `json` utility, but it does not seem to actually exist and there are a number of open issues asking for this feature, leading me to believe it isn't currently supported: https://docs.bazel.build/versions/master/skylark/lib/json.html

One hack I discovered is to wrap the value in a struct and use its built-in `to_json()` functionality with some string edits to get the inner value. It's a little awkward, but good enough for now.