Updates JavaScript example to use `inject_resources()`.
It's no longer necessary to hard-code a `<script />` tag, but instead we can rely on tooling to provide this line for the user.

For now, this macro isn't that useful as it replaces a single `<script />` line with a separate build macro without really gaining anything for the user. However this should grow over time to support styling and output a `web_resources()` rule which composes other `web_resources()` containing the actual JavaScript and CSS files.