Refactors effects to use a wrapper type instead of a `Proxy`.
The new `Effect` type is a simple wrapper of a dynamically retreived value. It replaces the old `Proxy` implementation because `Proxies` only really work with objects and do not play well with primitive types, meaning an effect of a `string` just doesn't work. Using a wrapper type has the unfortunate consequence of requiring users to call `.get()` in order to retrieve the value, however this involves much less magic as a result. It also means that the unique contract of effect values are reflected within the type system itself rather than being an awkward piece of documentation of the effect function.

This required refactoring a bunch of existing infrastructure to use the new `.get()` function and also updated `EffectTester` to align with the new API. Both of these are pretty simple changes that just applied to a lot of files in the repository, making this one of the larger commits.

Another benefit of this change is that we can drop the complicated `delegatedProxy()` implementation which is no longer needed.