Adds tests for `usePage()`.
I forgot to test this previously, so adding them now. This effect is unique in that it composes another effect (`Effect<Browser>`). This is the main use case for `effectFake()` to avoid spying issues that would arise from using two `EffectTester`s at the same time.

Also cleaned up a few of the mocks and assertions, since we can now assert on the object itself rather than an arbitrary property of it. This is one of the benefits of not using a proxy.