Adds `prerender_page_bundled()` macro.
This is a higher-level simplification of `prerender_page()`. The main advantage is that it handles bundling scripts for the user so they don't need to manually bundle themselves. It happens to use Rollup, but this is an implementation detail unimportant to the user.

The actual macro is pretty straightforward, simply rendering the page, bundling its scripts, and then injecting the script into the page. The final target is a `filegroup()` to encapsulate the two output files while everything else is considered an implementation detail.

I wanted to output the files at `%{name}.html` and `%{name}.js`, however if the user authors `%{name}.ts`, it will compile to `%{name}.js` and conflict. Since that is likely to be common, I figured it was simpler to just name the script `%{name}_bundle.js`.

Including the two files like this means referencing them from a devserver "just works", however they still have the name of the `prerender_page_bundled()` target, which the user may not want. For now, the example renames the HTML page to `index.html`, although it leaves the JavaScript because renaming it would be quite awkward. For now, this is just a little strange, but once we get `web_resources()` and let the user define a path to the page, this should get a lot cleaner.