Adds `styles` example.
This examples serves a minimal test case of CSS styling. It uses `postcss_binary()` to bundle CSS into a single file and then load it via a hard-coded `<link />` tag in the HTML document. It's pretty messy for the moment, but we can continue to iterate on it to handle as much as possible in the `rules_prerender` toolchain to shift the burden away from the application.

There is no officially supported or even well-known Bazel toolchain for CSS. There is one for Sass, but I don't want to require Sass with `rules_prerender`. PostCSS is a common transformation tool however and it can be tweaked to become a bundler, even though that's not its real goal. Until someone writes a real CSS ruleset for Bazel, this is the best we can do without inventing one from scratch.

`postcss_binary()` works on standard files, so we can use a simple `filegroup()` as a kind of "library" rule. These can compose like normal and are available to the bundler via the `deps` attribute.

The only actual plugin required to provide a bundler is `postcss-import`, which resolves `@import` statements and inlines their content. Given an entry point that imports all the dependencies, this is good enough to act as a bundler.

Unfortunately the configuration is surprisingly complicated. We need custom file resolution logic in order to load files correctly in a Bazel environment. This is further complicated by the fact that plugin configurations are loaded via a simple string which must be known at loading time, so we can even store this logic in a separate file for easy testing. For now it is housed as a string literal in a `*.bzl` file. We may want to revisit this by copying to a separate, testable file, and then have a test which verifies that the real implementation matches the test copy.

An entry point must be hand-authored atm and the bundled CSS must be manually included in the HTML document via a `<link />` tag. Both of these requires should be fixed as we implement the expected tooling.