Adds `styles` attribute to `prerender_*()` macros.
This allows users to define styles as files or `filegroup()`s and include them in `prerender_component()` which gets composed and propagated to the `prerender_page()` or `prerender_page_bundled()` macro at the top-level.

`prerender_page()` exports all the styles as `%{name}_styles`, while `prerender_page_bundled()` does **not** because it will automatically bundle these styles. This means the example is using the internals of `prerender_page_bundled()` somewhat, but this is temporary so its ok.