Adds `styles` property to `PrerenderMetadata`.
This will hold information about CSS styles. It currently *happens* to be identical to the JavaScript equivalent, but this is just chance and the two are likely to diverge over time.

Most of the work here is updating tests to not fail from introducing this value. I'm switching these tests to use mocks so they should be isolated from similar changes in the future.