Adds mocks for `PrerenderAnnotation` types.
The one unique aspect is the top-level type which is a discriminated union. To mock this, we just pick an arbitrary subtype (of which there is only one right now). It can be used as a placeholder when a test doesn't care at all about the actual annotation, only that one exists. However, `mockScriptAnnotation()` and future subtypes are likely to be more generally useful.