Adds `includeStyle()` calls to `styles` example.
This annotates the HTML with the files and gets extracted into the metadata file.

There is no automated test included here, but I manually confirmed that both the annotated HTML and `metadata.json` files hold the expected content. Once we generate a CSS entry point, this code path will be tested through that.