Bundles CSS in `prerender_page_bundled()`.
This adds a `postcss_binary()` rule to `prerender_page_bundled()` to bundle all the CSS on the page. It uses the styles `filegroup()` from `prerender_page()` with the auto-generated entry point as the target.

This allows the example app to no longer explicitly bundle the CSS, although it does still link to it manually. The next step is to inject the styles to remove the manual dependency and `<link />` tag.