Adds style support to resource injector.
This will accept styles with a path argument, read the contents of the path, and inject a `<style />` tag with the CSS content inlined.

We're using inlined styles here to avoid FOUC (flash of unstyled content) that comes with external stylesheets. Later we might add some options to lazy load certain styles, but for now all styles are inlined.