Updates `EMPTY_BUNDLE` warning to give a more actionable message.
When users generate an empty JavaScript bundle, the usual reason is because no JavaScript was included in the page. They now have the option to skip this build step, which will speed up builds, give one less file to serve, and improve performance by removing an unnecessary `<script />` tag.

This is left as a warning because it will still work, it's just a performance hit which gives no real gain.