Updates minimal example to use `prerender_page_bundled()`.
This skips bundling JavaScript and CSS because neither are included in this simplistic example.

An obvious question is, "If JS and CSS are not bundled, then why are we using `prerender_page_bundled()`?" This doesn't meaningfully change much at the moment, however as we add `web_resources()` this will aggregate all transitive resources with the page, which is still useful to be done automatically.