Adds `web_resources()` rule.
This provides a Bazel abstraction for aggregating a set of resources at specific paths. It mostly just calls through to the resource packager tool which does the heavy lifting.

The trickiest part here is allowing duplicate files. We want to support having the same file at multiple paths. Bazel rules support `attr.string_dict()` and `attr.label_keyed_string_dict()` options, but the former does not create a dependency while the later is keyed by labels and does not allow duplicates. To work around this, we use a `attr.string_dict()` which happens to have values which are the absolute paths of the relevant target. A separate `attr.label_list()` holds the de-duplicated labels and creates dependencies on them. We then replace the string value of the dictionary with its label equivalent to create the `string` -> `label` dictionary that we really want. This allows duplicate labels so the same file can be used at multiple paths.