Adds `ResourceMap.merge()`.
This merges multiple `ResourceMap` objects together into a single `ResourceMap`. Throws an error if there is a conflict in the URL path.

In the future we may allow a URL path conflict if the associated files have the same contents. For now, we'll use a more restrictive version and open it up later if appropriate.