Adds `deps` attribute to `web_resources()`.
This allows `web_resources()` rules to compose each other.

Introduced a new `_WebResourceInfo` provider to hold a structured reference to the resources directory. This could have been done via the existing reference as the only file in `DefaultInfo`, but using a distinct provider allows this to limit allowed `deps` to only other `web_resources()` rules.

The `_WebResourceInfo` provider is currently private, but we may consider exporting this in the future to interoperate with other tooling if necessary.