Adds `web_resources_devserver()`.
This creates a devserver and serves the content in the provided `web_resources()` target.

`ts_devserver()` is not quite what we need for this. `ts_devserver()` provides simple bundling of browser JavaScript which is unnecessary here, and I can't get `additional_root_paths` to serve the resources directory.

I opted to use `http-server` as a simplistic static file server implementation because I don't want to write and own my own for `rule_prerender`. This was wrapped in a Bash script to compile-in the resources directory so users can add arguments without changing the served directory. Doing so required a lot of fiddling with Bash runfiles. I'm pretty sure `$(rlocation)` and `$(rootpath)` are not designed to work together like this, but I was able to mangle the paths enough to work.

The primary downside of using an off-the-shelf server is that it is not Bazel-aware and does not support live reload. `ibazel run` seems to restart the server for some reason. I'm not entirely sure why but this is good enough for now.