Adds `web_resources_devserver()`.
This creates a devserver and serves the content in the provided `web_resources()` target. I was able to leverage `ts_devserver()` by abusing the `additional_root_paths` attribute. After a lot of debugging and attempts at other designs, I eventually discovered that adding a `data` dependency on a `web_resources()` target would not include its generated directory. It seems that I need to emit a `data_runfiles` value in the `DefaultInfo()` for the directory to be included.

Using `ts_devserver()` is a bit overkill, since its main purpose is to act as a simple bundler which we aren't using at all. However it does provide an existing devserver implementation with Bazel (and iBazel integration). The only awkward part is that live reload only works if you include `/_/ts_scripts.js` on the page. Normally that's not too big a deal, but currently we don't have a way of distinguishing between dev builds and prod builds, while this script should only be used on dev builds. We may consider adding some more tooling to inject this script into all HTML files in a `web_resources()` directory that gets loaded by a `ts_devserver()`.