Adds resources to `resources` example.
This introduces `web_resources()` rules which contain the resources for various components on the page. These example components use the resources as images at specified paths.

The same image content is reused out of laziness, but the source file is copied between the components and they load it at different paths. It is even loaded as a favicon (though this isn't easily testable with Puppeteer).

This uses the new `web_resources_devserver()` macro which acts mostly as a drop-in replacement.