Fixes injected JavaScript link in `prerender_page_bundled()`.
We were injecting the path to the bundled JavaScript, when the file would be moved to the path specified by the user, causing a broken `<script />` tag for users who provided a `path` attribute.

This updates the injected `<script />` tag to match with the final `web_resources()` target. Also updated the `scripts` example to use `web_resources_devserver()` because it would otherwise be broken by this change.