Updates examples to use `web_resources_devserver()` and be more consistent with each other.
This mostly involves replacing `ts_devserver()` with `web_resources_devserver()` and applying a `path` attribute so it ends up at the index page for simplicity. This allows us to remove an ugly `genrule()` from each example and leverage `web_resources()` in the custom bundling example.

I also removed a `<script />` tag which loaded the live reload script from `ts_devserver()`. While this technically still works as `web_resources_devserver()` is implemented via `ts_devserver()` (shh... don't tell anyone), it is not a part of the external contract and I cannot guarantee it will continue to work in future versions.