Adds `testonly` example.
This is a trivial example which just exercises as much code as it can while applying Bazel's `testonly` attribute to everything. This should act as a regression test to ensure that `testonly` is propagated to all the relevant places.

I opted to use a simple `build_test()` rather than a more complete Puppeteer-based test. Most of the actual functionality is tested by more specific examples and doesn't need to be repeated here. If there is a problem with `testonly`, the build will fail, so that's all we need to check.